test: fix four potency defects where the wave's own oracles could not fail
- the corrupt-checkpoint loop patched ONE shared file cumulatively, so
  the traversal/unicode resumeOf cases ran on a file still carrying the
  earlier attempt:2.5 poison — deleting the resumeOf validation left
  the gate green. Each case now patches a fresh copy (#1179)
- 'Cancel before dispatch' accepted ANY terminal state; a no-op
  requestCancel would have passed with an all-Succeeded run. The cancel
  cascade must end in Cancelled or Skipped only
- the IR/provenance mutation fuzz discarded both isSuccess() results:
  a mutation that made the parser ACCEPT garbage passed green. Acceptance
  now implies structural integrity (unique node ids, edges reference
  declared nodes)
- the data-scale structural-memory oracle SAID 64 KiB per record but
  enforced 64 MiB (~1000x the documented ceiling); enforce the
  documented bound with modest allocator slack

+++ tests/test_temporal_fit.cpp
@@ -460,8 +460,7 @@ TEST_CASE( "Seasonal decomposition trend is time-regularized, not sample-indexed
     {
         REQUIRE( std::isfinite( dense.trend[i] ) );
         REQUIRE( std::isfinite( coarse.trend[i] ) );
-        maxDelta = std::max( maxDelta,
-                             std::fabs( dense.trend[i] - coarse.trend[i] ) );
+        maxDelta = std::max( maxDelta, double( std::fabs( dense.trend[i] - coarse.trend[i] ) ) );
     }
     REQUIRE( maxDelta > 1e-4 );
 }
